requirements.txt
huggingface_hub
gradio
python-dotenv
rss-parser
tqdm
requests
numpy
feedparser
fastapi
uvicorn
gradio[mcp]